% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CALC-calc_binary.R
\name{calc_binary}
\alias{calc_binary}
\title{Calculate presence/absence.}
\usage{
calc_binary (tibble_ko, y_axis, data_experiment = NULL, binary = TRUE,
metabolism = FALSE)
}
\arguments{
\item{tibble_ko}{a tibble object from mapping_ko.}

\item{y_axis}{a character, indicating the pathway to analyze.}

\item{data_experiment}{optional. a data frame object
containing metadata information.}

\item{binary}{optional. a logical value, indicating whether you want to
calculate presence/absence or abundance of KOs per pathway.}

\item{metabolism}{optional. a logical value, if TRUE it will show
the metabolism table.}
}
\value{
A data frame containing the presence/absence (1/0) of each pathway
within each bin_name.
calc_binary(ko_bin_mapp, Pathway)
}
\description{
Calculate the presence/absence of data in certain pathway.
}
\details{
This function calculates the presence or absence of a pathway by
counting the number of times it appears (or not) in the dataset.
}
\examples{
# calc_binary (ko_bin_mapp, Pathway, binary = T, metabolism = T)
}
