% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{EC_picrust2}
\alias{EC_picrust2}
\title{EC abundances predicted with PICRUSt2}
\format{
A data frame with X rows and Y columns.
}
\source{
Generated from PICRUSt2.
}
\usage{
EC_picrust2
}
\description{
Dataset with EC number abundances predicted from functional profiling.
}
\keyword{datasets}
