% Generated by roxygen2: do not edit by hand
% Please edit documentation in r/2021_02_26-GET-get_pca.R
\name{get_pca_metabolism}
\alias{get_pca_metabolism}
\title{Obtains PCA data of KO pathways}
\usage{
get_pca_metabolism(ko_table)
}
\arguments{
\item{ko_table}{is a table object created with the get_unique function
and contains unique KOs.}
}
\description{
Identifies the most important KO pathways in each Bin.
}
\details{
This function is part of a package used for
the analysis of bins metabolism.
}
\examples{
get_pca_metabolism(Kegg_mapping)
}
