% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2021_02_25-GET-get_unique.R
\name{get_subset_unique}
\alias{get_subset_unique}
\title{Subsets the mapped KO table based on metadata to identify
KOs only present in a specific metadata group (i.e. KOs only present in Deltas and absent in the rest of database).}
\usage{
get_subset_unique(
  mapped_ko_table,
  other_data,
  type_of_interest_feature,
  interest_feature
)
}
\arguments{
\item{mapped_ko_table}{a data frame object. The output table from the mapping_ko function.}

\item{other_data}{a data frame object that contains the metadata (i.e. Taxonomy, sampling site).}

\item{type_of_interest_feature}{is a metadata column name. This feature is going to be used for subsetting.}

\item{interest_feature}{a string of the metadata feature of interest. It is found under the type_of_interest_feature value column.}
}
\description{
reads the output of the mapping_ko function to filter out KOs
based on metadata and identify KOs only present in a particular metadata group.
}
\details{
This function is part of a package used for the analysis of bins metabolism.
}
\examples{
get_subset_unique(ko_bin_mapp, metadata, Sample_site, "Water_column")
}
