% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2021_02_25-GET-get_unique.R
\name{get_unique}
\alias{get_unique}
\title{Identifies unique KOs in the data}
\usage{
get_unique(
  mapped_ko_table,
  other_data,
  type_of_interest_feature,
  interest_feature
)
}
\arguments{
\item{mapped_ko_table}{is the output table from the mapping_ko function.}

\item{other_data}{is a table that contains the necessary metadata.}

\item{type_of_interest_feature}{is the metadata column name containing different features.}

\item{interest_feature}{is a string of the metadata feature of interest found under the
type_of_interest_feature value column.}
}
\description{
reads the output of the mapping_ko function
to filter out KOs that are not mapped to a unique protein family.
}
\details{
This function is part of a package used for the analysis of bins metabolism.
}
\examples{
get_unique(ko_bin_mapp, metadata, Sample_site, "Water_column")
}
