% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{interpro_INTERPRO_profile}
\alias{interpro_INTERPRO_profile}
\title{interpro_INTERPRO_profile}
\format{
A  tiible of extra data of the bins.
}
\source{
\url{https://sites.utexas.edu/baker-lab/}
}
\usage{
interpro_INTERPRO_profile
}
\description{
The output of read_interpro for Interpro.
}
\keyword{datasets}
