% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{interpro_map}
\alias{interpro_map}
\title{interpro_map}
\format{
A tiible of extra data of the bins.
}
\source{
\url{https://sites.utexas.edu/baker-lab/}
}
\usage{
interpro_map
}
\description{
The output of mapping for KEGG from read_interpro.
}
\keyword{datasets}
