% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{interpro_pfam_profile}
\alias{interpro_pfam_profile}
\title{interpro_pfam_profile}
\format{
A  data frame of extra data of the bins.
}
\source{
\url{https://sites.utexas.edu/baker-lab/}
}
\usage{
interpro_pfam_profile
}
\description{
The output of read_interpro for PFAM.
}
\keyword{datasets}
