% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ko_bin_mapp}
\alias{ko_bin_mapp}
\title{Output of mapping_ko}
\format{
A data frame containing the pathway profile of six bins.
}
\source{
\url{https://sites.utexas.edu/baker-lab/}
}
\usage{
ko_bin_mapp
}
\description{
A dataset containing the output of  mapping_ko.
}
\keyword{datasets}
