% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ko_bin_table}
\alias{ko_bin_table}
\title{Output of read_ko}
\format{
A data frame with 6809 rows and four columns.
}
\source{
\url{https://sites.utexas.edu/baker-lab/}
}
\usage{
ko_bin_table
}
\description{
A dataset containing the output of read_ko.
}
\keyword{datasets}
