% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{KO_picrust2}
\alias{KO_picrust2}
\title{KO abundances predicted with PICRUSt2}
\format{
A data frame with X rows and Y columns (describe briefly).
}
\source{
Generated with PICRUSt2 using X sample set.
}
\usage{
KO_picrust2
}
\description{
This dataset contains KO (KEGG Orthology) predicted abundances from PICRUSt2 output.
}
\keyword{datasets}
