% Generated by roxygen2: do not edit by hand
% Please edit documentation in r/2021_02_25-MAPPING-mapping_ko.R
\name{mapping_ko}
\alias{mapping_ko}
\title{mapping_ko}
\usage{
mapping_ko(ko_abundance_table)
}
\arguments{
\item{ko_abundance_table}{an object created with read_ko.}
}
\description{
It reads an object created with read_ko and maps each KO
to the KEGG database.
}
\details{
This function is part of a package used for
the analysis of bins metabolism.
}
\examples{
mapping_ko(kegg_read)
}
