% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{metadata}
\alias{metadata}
\title{Metadata}
\format{
A  data frame of extra data of the bins.
}
\source{
\url{https://sites.utexas.edu/baker-lab/}
}
\usage{
metadata
}
\description{
A dataset containing the metadata of six bins.
}
\keyword{datasets}
