% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pathway_picrust2}
\alias{pathway_picrust2}
\title{Pathway-level predictions with PICRUSt2}
\format{
A data frame with X rows and Y columns.
}
\source{
PICRUSt2 results.
}
\usage{
pathway_picrust2
}
\description{
Pathway predictions from MetaCyc using PICRUSt2 output.
}
\keyword{datasets}
