% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2021_02_25-PLOT-plot_bubble_percentage.R
\name{plot_bubble_percentage}
\alias{plot_bubble_percentage}
\title{Visualize bin abundance as a percentage}
\usage{
plot_bubble_percentage(
  tabble_ko,
  other_data,
  x_axis,
  y_axis,
  size_bubble,
  metadata_feature
)
}
\arguments{
\item{tabble_ko}{is a table object from the get_pca_metabolism function
and contains the most important KO pathways in each Bin.}

\item{other_data}{is a table object containing metadata information.}

\item{x_axis}{is a string of the x axis label.}

\item{y_axis}{is a string of the y axis label.}

\item{size_bubble}{is a variable in tabble_ko specifying the size of bubbles in the bubble plot.}

\item{metadata_feature}{is a string of the column name in the metadata file.}
}
\description{
Creates a bubble plot of each Bin's abundance by percentage.
}
\details{
This function is part of a package used for the analysis of bins metabolism.
}
\examples{
plot_bubble_percentage(ko_bin_mapp, metadata, Bin_name, Pathway, Percentage, Clades)
}
