% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PLOT-disc_effect.R
\name{plot_disc_effect}
\alias{plot_disc_effect}
\title{Plot differential effect for discriminant features}
\usage{
plot_disc_effect(disc_obj, top_n = 40)
}
\arguments{
\item{disc_obj}{Object stored as attr(, "rbims_disc") from get_subset_discriminant().}

\item{top_n}{Number of features to display.}
}
\value{
A ggplot object.
}
\description{
This plot shows the effect size (x-axis) for top discriminant features.
Point size reflects significance as -log10(FDR), and color encodes effect direction.
}
