% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rbims}
\alias{rbims}
\title{Main Rbims}
\format{
Home made database of intresting pathways
}
\source{
\url{https://sites.utexas.edu/baker-lab/}
}
\usage{
rbims
}
\description{
Main Rbims
}
\keyword{datasets}
