% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2022_08_15-READ-read_dbCAN3.R
\name{read_dbcan3}
\alias{read_dbcan3}
\title{Read the output of dbCAN and extract abundance profile}
\usage{
read_dbcan3(dbcan_path, write=FALSE, profile=TRUE)
}
\arguments{
\item{dbcan_path}{a path where dbCAN output data are. They
should have the extension overview.txt and all files in the path are the ones that
need to be read. Output data should have 6 columns with the bin names
followed by the Genes obtained in every algorithm (HMMER,Hotpep,DIAMOND),
column 'Signalp' indcating if a Peptide signal is found and a column
'#ofTools" indicating the number of algorithms that found this Gene.}

\item{write}{a logical value indicating to save the data imported
as a formatted table with .tsv extension with a time stamp and it will be
located in your current workin directory}

\item{profile}{a logical value indicating if you want to print a profile
or not.}
}
\description{
read_dbcan3 calculates the abundance of each Gene within the
bins based on the dbCAN output from run_dbcan3 V3.0.6.
}
\examples{
\dontrun{
read_dbcan3("C:/Users/bins/")
}
}
