% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/READ-read_interpro.R
\name{read_interpro}
\alias{read_interpro}
\title{Extract abundance profile of InterProScan output.}
\usage{
read_interpro (data_interpro = NULL, 
database = c("KEGG", "Pfam", "INTERPRO",
"TIGRFAM", "SUPERFAMILY", "SMART", "SFLD", "ProSiteProfiles",
"ProSitePatterns", "ProDom", "PRINTS", "PIRSF", 
"MobiDBLite","Hamap", "Gene3D", "Coils", "CDD"), profile = TRUE, write = FALSE)
}
\arguments{
\item{data_interpro}{a table, output of InterProScan on tsv format.
InterProScan should have been run with -pa option to be able to use the
KEGG option, in the database argument.}

\item{database}{a character indicating for which database do you want to
get the abundance profile. Valid options are "KEGG", "PFAM" or "INTERPRO".}

\item{profile}{a logical value indicating if you want to print a profile
or not. This option is valid for "PFAM" and "INTERPRO" database.}

\item{write}{a logical value indicating to save the data imported
as a formatted table with .tsv extension with a time stamp and it will be
located in your current working directory}
}
\description{
Reads a table object created with InterProScan and generates
a profile table of abundance with the hits of the KEGG, Pfam or
INTERPRO databases. The output of KEGG database can be used within
mapping_ko.
}
\details{
This function is part of a package used for the analysis of bins
metabolism.
}
\examples{
filepath <- system.file("extdata", "Interpro_test.tsv", package = "rbims")
read_interpro(data_interpro = filepath, database = "INTERPRO", profile = FALSE)
}
