% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/READ-read_ko.R
\name{read_ko}
\alias{read_ko}
\title{Read the output of
KofamScan/\href{https://www.genome.jp/tools/kofamkoala/}{KofamKoala} or
\href{https://www.genome.jp/kegg/kaas/}{KAAS}.}
\usage{
read_ko(data_kofam=NULL, data_kaas=NULL, data_interpro=NULL, 
write=FALSE)
}
\arguments{
\item{data_kofam}{a path where KofamScan/KofamKoala output data are. They
should have the extension .txt and all files in the path are the ones that
need to be read. Output data should have 5 columns with the bin names
followed by the scaffold name divided by a '-' or '_': bin_scaffoldXX.}

\item{data_kaas}{a data frame with 2 columns. Contigs are expected to
indicate in their names the bin name followed by the scaffold name
divided by a '-' or '_': bin_scaffoldXX.}

\item{data_interpro}{a data frame output of read_interpro. This
argument is used within mapping_ko.}

\item{write}{a logical value indicating to save the data imported
as a formatted table with .tsv extension with a time stamp and it will be
located in your current working directory}
}
\description{
read_ko calculates the abundance of each KO within the
bins based on the KofamScan or KofamKoala output.
}
\details{
This function is part of a package used for the analysis
of bins metabolism.
}
\examples{
kofam_dir <- system.file("extdata", "KEGG_bins", package = "rbims")
read_ko(data_kofam = kofam_dir, write = FALSE)
}
