% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2022_04_07-READ-read_ko_shiny.R
\name{read_ko_shiny}
\alias{read_ko_shiny}
\title{Read the output of KofamScan/KofamKoala or KAAS.}
\usage{
read_ko_shiny(data_kofam=NULL, data_kaas=NULL, data_interpro=NULL)
}
\arguments{
\item{data_kofam}{a file output of KofamScan/KofamKoala. They
should have the extension .txt and all files in the path are the ones that
need to be read. Output data should have 5 columns with the bin names
followed by the scaffold name divided by a '-' or '_': bin_scaffoldXX.}

\item{data_kaas}{a data frame with 2 columns. Contigs are expected to
indicate in their names the bin name followed by the scaffold name
divided by a '-' or '_': bin_scaffoldXX.}

\item{data_interpro}{a data frame output of read_interpro. This
argument is used within mapping_KO.}
}
\description{
read_ko_shiny calculates the abundance of each KO within the
bins based on the KofamScan or KofamKoala output.
}
\details{
This function is part of a package used for the analysis
of bins metabolism.
}
\examples{
\dontrun{
read_ko_shiny("Kofam_output")
}
}
