% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/READ-read_picrust2.R
\name{read_picrust2}
\alias{read_picrust2}
\title{Read the output of PICRUSt2}
\usage{
read_picrust2(data_picrust2, profile=TRUE, write=FALSE,
database= c("KO", "EC", "pathway"))
}
\arguments{
\item{data_picrust2}{a table, output of PICRUSt2 on tsv format.}

\item{profile}{a logical value indicating if you want to print a profile
or not.}

\item{write}{a logical value indicating to save the data imported
as a formatted table with .tsv extension with a time stamp}

\item{database}{a character indicating for which database do you want to
get the abundance profile. Valid options are "KO", "EC" or "pathway".}
}
\description{
read a table created in PICRUSt2 and gives format to downstream analysis
}
\details{
This function is part of a package used for the analysis of bins
metabolism.
}
\examples{
\dontrun{
read_picrust2(data_picrust2="inst/extdata/pred_metagenome_unstrat.tsv",
database="KO", profile = F, write=F)
}
}
