% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2021_10_22-WRITE-write_metabolism.R
\name{write_metabolism}
\alias{write_metabolism}
\title{Metabolism table}
\usage{
write_metabolism(data_interpro=NULL, data_ko=NULL)
}
\arguments{
\item{data_interpro}{a table, output of InterProScan on tsv format.
InterProScan should have been run with -pa option to be able to use the
KEGG option, in the database argument.}

\item{data_ko}{path to the directory where the KEGG output files are.}
}
\description{
Creates a metabolism table that includes the information from
KEGG and Interpro annotation.
}
\details{
This function is part of a package used for the analysis of bins
metabolism.
}
\examples{
\dontrun{
write_metabolism("inst/extdata/Interpro_test.tsv", 
"inst/extdata/")
}
}
