% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2021_04_28-CALC-calc_percentage.R
\name{calc_percentage}
\alias{calc_percentage}
\title{Calculate the percentage.}
\usage{
calc_percentage(tibble_ko, y_axis, data_experiment=NULL)
}
\arguments{
\item{tibble_ko}{a tibble object from mapping_ko.}

\item{y_axis}{a character, indicating the pathway to analyze.}

\item{data_experiment}{optional. a data frame object
containing metadata information.}
}
\description{
Calculate the percentage of KO in certain pathway.
}
\details{
This function is part of a package used for the analysis of bins
metabolism.
}
\examples{
calc_percentage(ko_bin_mapp, Pathway)    
}
