% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2021_02_26-GET-get_pca.R
\name{get_subset_pca}
\alias{get_subset_pca}
\title{Obtains PCA data of KO pathways.}
\usage{
get_subset_pca(ko_table)
}
\arguments{
\item{ko_table}{a data frame object, created with the mapping_ko or get_subset_* functions.}
}
\description{
Identifies the most important KO pathways in the whole databse.
}
\details{
This function is part of a package used for
the analysis of bins metabolism.
}
\examples{
get_subset_pca(ko_bin_mapp)
}
