% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2021_02_25-GET-get_unique.R
\name{get_subset_unique}
\alias{get_subset_unique}
\title{Unique abundance profile.}
\usage{
get_subset_unique(tibble_rbims, data_experiment, experiment_col, 
experiment_col_element, analysis=c("KEGG", "PFAM", "INTERPRO"))
}
\arguments{
\item{tibble_rbims}{a tibble object, created with
the read_interpro or mapping_ko functions.}

\item{data_experiment}{a data frame object that contains the metadata
(i.e. taxonomy, sampling site).}

\item{experiment_col}{a metadata column name.
This feature is going to be used for sub-setting.}

\item{experiment_col_element}{a string of the metadata feature of interest.
It is found under the experiment_col column.}

\item{analysis}{a character, indicating from which input do you want to
get the unique abundance profile. Valid options are "KEGG", "PFAM"
or "INTERPRO".}
}
\description{
reads the output of the read_interpro or mapping_ko functions
to pull out the unique IDs based on metadata. This function identifies
the IDs that are only present in a particular metadata group (i.e. KOs only
present in environment "a" and absent in the rest of the environments).
And retrieves a tibble object.
}
\details{
This function is part of a package used for the analysis
of bins metabolism.
}
\examples{
get_subset_unique(ko_bin_mapp, metadata, Sample_site, 
"Water_column", analysis="KEGG")
}
