% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2021_04_28-PLOT-plot_bubble.R
\name{plot_bubble}
\alias{plot_bubble}
\title{Bubble plot.}
\usage{
plot_bubble(tibble_ko, x_axis, y_axis, analysis=c("KEGG","INTERPRO", "DBCAN"),
data_experiment=NULL, calc=NULL, color_character=NULL, order_bins=NULL, 
order_metabolism=NULL, color_pallet=NULL, range_size=NULL, x_labs=TRUE,
y_labs=TRUE, text_x=NULL, text_y=NULL)
}
\arguments{
\item{tibble_ko}{a tibble object, created with the read_*
mapping_ko and get_subset_* functions.}

\item{x_axis}{a string, a column name of the metabolism table.
It determined the x axis label.}

\item{y_axis}{a string, a column name of the metabolism table.
It determined the y axis label.}

\item{analysis}{a character indicating if your input data are from
KEGG, INTERPRO or DBCAN}

\item{data_experiment}{optional. a data frame object
containing metadata information.}

\item{calc}{a character indicating with type of calc should
be done to plot the results. Valid values are "Abundance", "Binary",
"Percentage", and "None". If you chose none you are expected to use a
tibble table obtained from calc_binary or calc_percentage.}

\item{color_character}{optional. a string column name of the metadata
or metabolism object, used for color.}

\item{order_bins}{optional. a character vector indicating the bin order.}

\item{order_metabolism}{optional. a character vector
indicating metabolism order.}

\item{color_pallet}{optional. a character vector of colors to use.}

\item{range_size}{optional. a numeric vector indicating
the range size of the dots.}

\item{x_labs}{optional. If FALSE it will set the x lab to NULL.}

\item{y_labs}{optional. If FALSE it will set the y lab to NULL.}

\item{text_x}{optional. A numeric vector indicating the size
of the x text letters.}

\item{text_y}{optional. A numeric vector indicating the size
of the y text letters.}
}
\description{
Creates a bubble plot of the data of bins metabolism
}
\details{
This function is part of a package used for
the analysis of bins metabolism.
}
\examples{
plot_bubble(tibble_ko=ko_bin_mapp, x_axis=Bin_name, y_axis=Module, 
analysis="KEGG", data_experiment=metadata, calc="Binary",
color_character=Clades)
}
