% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2021_02_25-PLOT-plot_bubble_abundace.R
\name{plot_bubble_abundance}
\alias{plot_bubble_abundance}
\title{Bubble plot of KO/Pathways/Modules abundance within each Bin.}
\usage{
plot_bubble_abundance(tabble_ko, other_data, x_axis, y_axis, metadata_feature)
}
\arguments{
\item{tabble_ko}{a data frame object, created with the mapping_ko or get_subset_* functions.}

\item{other_data}{a data frame object containing metadata information.}

\item{x_axis}{a string, it determined the x axis label.}

\item{y_axis}{a string, it determined the y axis label.}

\item{metadata_feature}{a string column name of the metadata object, used for color.}
}
\description{
Creates a bubble plot of KO/Pathways/Modules abundance within each bin. It uses the metadata information to color bubbles.
}
\details{
This function is part of a package used for
the analysis of bins metabolism.
}
\examples{
plot_bubble_abundance(ko_bin_mapp, metadata, Bin_name, Pathway, Clades)   
}
