% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2021_02_25-PLOT-plot_heatmap_abundace.R
\name{plot_heatmap_abundance}
\alias{plot_heatmap_abundance}
\title{Visualize bin abundance as a heatmap}
\usage{
plot_heatmap_abundance(tabble_ko, other_data, plot_ano, plot_medata)
}
\arguments{
\item{tabble_ko}{is a table object from the get_pca_metabolism function
and contains the most important KO pathways in each Bin.}

\item{other_data}{is a table object containing metadata information.}

\item{plot_ano}{is a column name of the KO module of interest in tabble_ko.}

\item{plot_medata}{is the column of interest in the metadata file.}
}
\description{
Creates a heatmap of each Bin's abundance.
}
\details{
This function is part of a package used for the analysis of bins metabolism.
}
\examples{
plot_heatmap_abundance(ko_bin_mapp, metadata, Module, Sample_site)
}
