% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2021_03_04-PLOT-plot_heatmap_porcentage.R
\name{plot_heatmap_percentage}
\alias{plot_heatmap_percentage}
\title{Heatmap plot of KO/Pathways/Modules and its relative
abundance within each bin.}
\usage{
plot_heatmap_percentage(
  tibble_ko,
  data_experiment,
  metabolism_col_sort,
  metabolism_col,
  experiment_col
)
}
\arguments{
\item{tibble_ko}{a data frame object, created with the mapping_ko
or get_subset_* functions.}

\item{data_experiment}{a data frame object containing metadata information.}

\item{metabolism_col_sort}{a string. A column name of the tibble_ko of a
feature to plot (i.e. KO/Pathways/Modules).}

\item{metabolism_col}{is a column name indicating the annotation if rows.
This column name comes from the mapping_ko or get_subset_* object.}

\item{experiment_col}{is a column name indicating the annotation if cols.
This column name comes from the metadata object.}
}
\description{
Creates a heatmap of KO/Pathways/Modules and
its relative abundance within each Bin.
}
\details{
This function is part of a package used for
the analysis of bins metabolism.
}
\examples{
plot_heatmap_percentage(ko_bin_mapp, metadata, Pathway, Module, Clades) 
}
