% Generated by roxygen2: do not edit by hand
% Please edit documentation in r/2021_02_25-READ-read_ko.R
\name{read_ko}
\alias{read_ko}
\title{read_ko read the output of KofamScan or KofamKoala}
\usage{
read_ko(ko_result)
}
\arguments{
\item{ko_result}{KEGG output file. Contings are expected
to be named similar to bin_scaffoldXX.}
}
\description{
read the output of KofamScan or KofamKoala and
calculate the abundance of each KO within the Bins.
}
\details{
This function is part of a package used for
the analysis of bins metabolism.
}
\examples{
read_ko("KEGG_bins.txt")
}
